requirements.txt
alembic==1.8.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==22.1.0
bcrypt==4.0.1
blinker==1.5
cachelib==0.9.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
coverage==6.5.0
defusedxml==0.7.1
Deprecated==1.2.13
exceptiongroup==1.0.4
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Caching==2.0.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
flask-restx==1.0.3
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
httpie==3.2.1
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.0
Mako==1.2.3
MarkupSafe==2.1.1
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
pytest==7.2.0
pytest-cov==4.0.0
python-dotenv==0.21.0
pytz==2022.6
redis==4.3.4
requests==2.28.1
requests-toolbelt==0.10.1
rich==12.6.0
SQLAlchemy==1.4.43
tomli==2.0.1
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
